# MemoryKnowledge/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge Service API
  description: |
    Standalone knowledge service for Code-Graph + LLM-Wiki.
    Provides endpoints for managing and querying code repositories
    and wiki knowledge bases.

    ## Multi-tenancy (001)

    Every endpoint requires the tenant identity `service_id` in the
    **`x-tdai-service-id` request header** (`^[A-Za-z0-9_-]+$`, missing/malformed → 400).
    This is the same value the kernel uses for store/VDB routing, so tenant identity is
    unified end-to-end. All lookups are scoped by `service_id`; an id-only endpoint
    targeting another tenant's resource returns 404. `service_id` is never a request-body
    field and is not echoed in responses.

    ## Asset ID Model

    Assets have globally unique, immutable IDs:

    | Asset Type  | Field           | Prefix   | Example        |
    |-------------|-----------------|----------|----------------|
    | LLM-Wiki    | `wiki_id`       | `wiki-`  | `wiki-9c1f2b`  |
    | Code-Graph  | `code_graph_id` | `cg-`    | `cg-7e3d10`    |

    ## Response Envelope

    All responses use `ApiResponseEnvelope`:
    - `code: 0` means success; non-zero means business error
    - `data` contains the payload on success, `null` on error
    - `message` provides human-readable status

  version: "1.0.0"
  contact:
    name: Knowledge Service
servers:
  - url: http://localhost:8421/v3
    description: Local development server
...
paths:
...
  /source-credential/put:
    post:
      tags: [Code-Graph]
      summary: Create or rotate an owner-only Git credential. Secret fields are write-only.
      security:
        - GitServiceBearer: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GitCredentialPutRequest'
      responses:
        '200':
          description: Success envelope; secrets are never returned.
        '400':
          description: Invalid request or inaccessible repository.
        '401':
          description: Service authentication required.
        '404':
          description: Credential not found in the service/team/owner scope.
        '409':
          description: Credential is still referenced by a code graph.
        '503':
          description: Service authentication or encryption key is not configured.
...
  /source-credential/delete:
    post:
      tags: [Code-Graph]
      summary: Delete an unused Git credential. Returns 409 if a graph still references it.
      security:
        - GitServiceBearer: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GitCredentialScopeRequest'
      responses:
        '200':
          description: Success envelope; secrets are never returned.
        '400':
          description: Invalid request or inaccessible repository.
        '401':
          description: Service authentication required.
        '404':
          description: Credential not found in the service/team/owner scope.
        '409':
          description: Credential is still referenced by a code graph.
        '503':
          description: Service authentication or encryption key is not configured.
...
components:
  securitySchemes:
    GitServiceBearer:
      type: http
      scheme: bearer
      description: Knowledge service key. Required even in legacy open mode for Git credentials.
  parameters:
    ServiceIdHeader:
      name: x-tdai-service-id
      in: header
      required: true
      description: >-
        Tenant identity `service_id` (001 multi-tenancy). Required on every endpoint;
        same value the kernel uses for store/VDB routing. Missing/malformed → 400.
      schema:
        type: string
        pattern: "^[A-Za-z0-9_-]+$"
...
  schemas:
    GitCredentialScopeRequest:
      type: object
      required: [team_id, user_id]
      properties:
        team_id:
          type: string
        user_id:
          type: string
          description: Authenticated user identity asserted by Panel, not an end-user authorization token.
        credential_id:
          type: string
          description: Required by test/delete; omit for list.
...
    GitCredentialPutRequest:
      allOf:
        - $ref: '#/components/schemas/GitCredentialScopeRequest'
        - type: object
          required: [name, secret]
          properties:
            name:
              type: string
              maxLength: 128
            hostname:
              type: string
              example: cnb.cool
              description: Required for HTTPS. Hostname only, without protocol, port, user credentials or path. Omit for SSH.
            secret:
              writeOnly: true
              oneOf:
                - type: object
                  required: [kind, username, token]
                  properties:
                    kind:
                      type: string
                      enum: [https]
                    username:
                      type: string
                    token:
                      type: string
                      writeOnly: true
                - type: object
                  required: [kind, private_key]
                  properties:
                    kind:
                      type: string
                      enum: [ssh]
                    private_key:
                      type: string
                      writeOnly: true
...